fgi/image: ignore error and warning when webp converting failed

+++ fgi/image.py
@@ -166,13 +166,18 @@ def uri_to_html_image(rr, imageuri, gameid, alt = None):
     if args.args.images_to_webp \
             and not img.is_remote \
             and webp.can_convert(path):
-        img.uri += ".webp"
-        img.path += ".webp"
 
-        if os.path.exists(path):
-            if not os.path.exists(img.path):
-                webp.cwebp(path, img.path)
-            os.remove(path)
+        new_path = img.path + ".webp"
+
+        if os.path.exists(path) \
+                and not os.path.exists(new_path):
+            try:
+                webp.cwebp(path, new_path)
+                os.remove(path)
+                img.uri += ".webp"
+                img.path = new_path
+            except:
+                print(f"[warning] {path} can not be converted to webp")
 
     hi = HTMLImage.from_image(img)
     hi.alt = alt
